TransferPolicy created. access and create methods added to it. access authorization in TransferController@destroy. Gate transfer-to-self defined in AuthServiceProvider and registered there. 4 Authorizations for TransferController@create added. 1 Authorization added to TaskController@destroy.

+++ app/Http/Controllers/TaskController.php
@@ -40,6 +40,10 @@ class TaskController extends Controller
     public function destroy(Request $request, Task $task)
     {
       $this->authorize('destroy', $task);
+
+      // when a task is IN TRANSFER it CANNOT BE DELETED.
+      abort_if($task->transfer, 403);
+
       $task->delete();
 
       return redirect('/tasks');
@@ -47,7 +51,7 @@ class TaskController extends Controller
 
     public function show(Task $task)
     {
-      
+
     }
 
 }
